requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Cython==0.29.14
Django==2.2.7
h5py==2.10.0
idna==2.8
image==1.5.27
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pycocotools==2.0.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
torch==1.3.1
torchvision==0.4.2
urllib3==1.25.7
